+++ src/benchmark.cpp
@@ -6,6 +6,7 @@
  */
 
 #include <fmt/color.h>
+
 #include <taskbench/benchmark.h>
 #include <taskbench/utils/statistics.h>
 #include <taskbench/utils/format.h>
@@ -13,7 +14,7 @@
 namespace taskbench {
 
 // _____________________________________________________________________________________________________________________
-BenchmarkResult::BenchmarkResult(std::string name, size_t data_size, size_t num_operations)
+BenchmarkResult::BenchmarkResult(std::string name, uint64_t data_size, uint64_t num_operations)
     : _name(std::move(name)), _data_size(data_size), _num_operations(num_operations) {}
 
 // _____________________________________________________________________________________________________________________
@@ -79,7 +80,7 @@ nlohmann::json BenchmarkResult::json() {
 const std::string& BenchmarkResult::name() const { return _name; }
 
 // _____________________________________________________________________________________________________________________
-size_t BenchmarkResult::data_size() const { return _data_size; }
+uint64_t BenchmarkResult::data_size() const { return _data_size; }
 
 // === AbstractBenchmark ===============================================================================================
 // _____________________________________________________________________________________________________________________
@@ -92,7 +93,7 @@ std::map<std::string, BenchmarkResult> AbstractBenchmark::results() { return _be
 const std::map<std::string, BenchmarkResult>& AbstractBenchmark::results() const { return _benchmark_result; }
 
 // _____________________________________________________________________________________________________________________
-void AbstractBenchmark::_register_benchmark(size_t data_size, size_t num_operations, const std::string& name) {
+void AbstractBenchmark::_register_benchmark(uint64_t data_size, uint64_t num_operations, const std::string& name) {
   _benchmark_result.insert({name, BenchmarkResult(name, data_size, num_operations)});
 }
 // _____________________________________________________________________________________________________________________
